feat: Update Backup and Restore documentation
The Backup and Restore documentation has been added to provide users with instructions on how to backup their data and restore it in case of data loss. The documentation covers topics such as backing up the home directory, dconf settings, and the /etc directory. It also includes information on automating the backup process using a cron job.

Fixes #456