feat: Update Linux kernel note in Check-File-Format.md
This commit updates the note about the Linux kernel being an executable file in the Portable Executable (PE) format in the "Check-File-Format.md" file. It clarifies that the Linux kernel uses the PE format to allow it to be booted directly by UEFI firmware on compatible systems. This improves the accuracy and completeness of the documentation.

Note: Please remove any meta information like issue references, tags, or author names from the commit message.