```text chore: Update /etc/hosts with new hostname
This commit updates the Linux.md file to include instructions for updating the /etc/hosts file with the new hostname. This is necessary to ensure that the new hostname is associated with the correct IP address. The instructions provide the necessary steps to add the new hostname as 127.0.1.1 in the /etc/hosts file.